move browse name class to application directory
